app/requirements/local.txt
-r ./base.txt

click==8.3.1
django-environ==0.12.0
pre-commit==4.5.1

pytest==9.0.2
redis==7.4.0
